refactor(common): establish currency domain owner
